Refactor MscmpSystPerms database functional design
The prior design of the permissioning system was more complex than it
had to be (good argument that it still is) and didn't fit all of the
potential use cases either.  Part of this was that it was too
configurable on an ad hoc basis; this was especially apparent in the
idea behind fully normally representation of access degrees.  The degree
system feels like it would have been both complicating and slow to use
in practice because of the increased normalization in the database.

The new design gets rid of the degree concept and the type concept which
was part of it.  It ultimately brings a different conception of type,
but not one which is user configurable and enables arbitrary kinds of
permissions.  The new type is simply a grouping mechanism for applying
permission roles and permissions to different application contexts.  The
new design standardizes on four "Rights" which apply to what data can be
created/deleted, maintained, and viewed as well as standardized &
limited possible "Scopes" of applicability (such as "see own" or "see
all" kinds of scope).  While permissions can designate which specific
scopes are supported for each right, they still must designate from the
standard set.  We continue to support the notion of custom permissions,
but not custom permission functional types which are all system defined
and directly relatable to functional areas of any given application.

This current design feels more flexible, allows an application to offer
different permissions in different contexts, and balances cohesive
functionality with flexibility.  Time will naturally tell if the current
good feelings for the new design are warranted, but it feels like the
result of better thinking than the prior effort...  which never quite
felt comfortable.