Fix syst_perm_role_grants trigger context query
There were two problems in this query.

The larger problem was that we were trying to look at the assigned
scopes via the syst_perm_roles record... but that record doesn't contain
any scope references.  What we really needed to be doing was comparing
the "new" record scope values with data retrieved in part of the context
query.

The next problem was simply a misspelled column reference.