Refactor MscmpSystPerms.Types.rights_scope/0
Change the type from String.t() to an enumeration of atoms.  It really
should have been this from the start since we have a fixed number of
predefined values.

Also, update the Permissions & Permission Role Grant params to expect
atoms rather than strings (as well as their tests).  This isn't perfect:
instances of Msdata.SystPerms and Msdata.SystPermRoleGrants will still
want to work in strings.  Handling Atoms in the changeset functions for
inserts and updates is simple enough, the return values from queries
retrieving data from the database is less good.  We can define the
database fields to handle either strings or atoms, but my feeling is
that we should clearly want one or the other.

A custom database type may be in order; an "AtomString" or "DbAtom" or
some such thing where we expect both the input and final output to be
Atoms.  I can understand why no such standard Postgrex/Ecto type exists,
especially given the constraints around converting arbitrary strings to
Atoms, but in this case it would be helpful.  Something to think about
if this gets annoying.