Fix extraneous MscmpSystPerms user_description req
In the update_changeset of several record types, we require the
user_description field, but in the update context the value is optional
and may even be explicitly nil.

We should continue to require a non-nil value in this field for insert
changesets; user created records should include a short description.
Yes, there's an inconsistency here, but it's not an urgent one.  We'll
finesse it later.