Refactor MscmpSystPerms compare_scopes/2
Previously compare_scopes/2 was a private function in the
SystPermRoleGrants validators, but it's likely to be more generally
useful.

Move compare_scopes/2 to be a public function in Impl.PermRoleGrant and
also add it to the Public API.  Also generalize it to accept either
Scopes as Atoms or Scopes as Strings.