Refactor Ecto validator constraint captures
A fair number of our changeset functions weren't really doing anything
to capture constraint violations.  This commit corrects that.

In a few places for internal_name and display_name we had cross-record
unique_constraint/3 calls that might have worked by coincidence but
were not guaranteed to since we're naming our own constraints in the DB
migrations.  This has been resolved.

Also, fix a constraint name in syst_access_account_instance_assocs.  The
name was based on the prior definition of the constraint, but the
columns involved were changed and so the name should be updated.