Refactor diagnostic column default loading
Currently pretty much all the data tables include diagnostic columns.
Some of these are typical and found in many applications and some of
them are more... exotic.  However, a number of them will not be so
useful in the application and to users; they do provide diagnostics for
those systems administrators & developers that might investigate certain
kinds of issues or runtime characteristics.

This change flips the default query behavior of the application from
defaulting to load this usually irrelevant data to not loading it.
Hopefully we'll save a little database resource consumption not bringing
this stuff over all the time.