Add perms ms_syst_priv.get_greatest_rights_scope
This DB function is intended to help resolve the effective Scope for a
Permission Grant.  A user may be granted multiple Permission Roles which
individually can grant access to the same Permission with different
Scopes for its rights.  Our rule is that the grant with the broadest/
greatest/most expansive Scope wins.

This function takes an array of Scopes presumably from the same
Permission and will return the broadest/etc Scope.  If the input is
empty, null, or otherwise unresolvable the return Scope of last resort
is `deny`.