Add additional MscmpSystPerms.Protocol functions
This includes additional Protocol function specifications as well as an
API for accessing these in a reasonable generic way (not unlike Enum).

Note that there are still functions will which need to be added to the
protocol (e.g. deny_perm/2 is one I'm pretty sure we'll need).  But
we'll wait until we see that need before jumping on it.