Refactor CI conf MuseBMS to MssubBms/MsappPlatform
This starts to sort out the creation of MsappPlatform and the turning of
BMS into a subsystem in regard to CI testing.  

The expectation is that these changes will blow up on the next CI run, 
but that we'll get some feedback about what needs to be done to get 
everything working.