Refactor Platform
Where to even begin.  Working out how to handle both a administrative
UI (MCP) and the public API (BMS for now), it became apparent that the
best thing to do was to encapsulate the web front ends and service
related stuff in a Phoenix oriented umbrella application.  This
conclusion was also reached after some discussion on the Elixir Forums
(https://elixirforum.com/t/phoenix-1-7-rc0-project-organization-multi-endpoint-ui/52968/1)

This effectively makes decisions about what "platform" is about.  The
subsystems and components will continue the previous development pattern
for now.  It may make sense to get rid of mssub_mcp and just make it in
msplatform/apps/msapp_mcp.

There's still a fair amount of thinking to go into the implications of
the path Phoenix and its generators are taking us down, but at least we
have a start.