Add msplatform_state checking logic to MsappMcp
Adds some basic logic to check if we should be in "bootstrap mode" or
not.

This commit raises some important structural and approach questions.
The lookup of the `msplatform_state` setting directly calls the
MscmpSystSettings API rather than calling through some MsssubMcp
wrapper.  This is a bit of change of direction from other calls which
rely do rely on the MssubMcp wrapper API.  Previously, I've discussed
that the MssubMcp wrapper is just unnecessary abstraction, but on the
other hand writing this Msplatform code felt like it would have been
more natural to call a MssubMcp wrapper than to hit the lower level
component more directly.  This is because this is the only place where
I'm having to worry about managing which services need to be active.

For now we'll go with this split solution... but I expect that my
misgivings about the wrapper will ebb some more and we'll likely move
this setting lookup to MssubMcp.