Add starting bootstrap web page
This is pretty early and probably not well constructed, but it's a
start.

We add the first routing which says if the system is configured for
bootstrap mode (the system default) to go ahead and redirect to the
platform bootstrap wizard.  Currently this works, though I expect that
structurally the code isn't where the code belongs... we'll see how my
feelings on that topic evolve while writing more of the logic.

At this commit we have a very basic static page wireframish kind of
layout and some placeholders.  Some care was given to spacing and such.
After this commit we'll start looking at how to make this dynamic and
interactive with our form data.