Refactor MscmpSystDb privileged DB role handling
Previously, the privileged database role `ms_syst_dba` was a manual
script created role with PostgreSQL CREATEDB and CREATEROLE privileges.
This role was used for the creation of Instance databases and the
backing context roles on which they depended.  While `ms_syst_dba` was
not a postgres SUPERUSER, it was pretty close and had the power to
arbitrarily create PostgreSQL superuser accounts.  Naturally, such an
account being able to be used by the application to connect to the
database considerably weakened security around our model; ironic since
the model was intended to help us make things more secure.

On reflection, I realized that we didn't need to have the application
login account which could create databases and database roles be granted
the CREATEROLE privilege.  Instead we could create a non-login role
which had this privilege and then create SECURITY DEFINER functions
which could then wrap very specific kinds of CREATE ROLE/DROP ROLE
statements which could then be run by the application connection role.
More than just preventing arbitrary CREATE ROLE commands this way, we
can designate that only those roles created using this process can be
deleted by the automated process.

This commit implements the mechanics to make this happen.  Along the way
we reduce the number of duplicate bootstrap_database scripts which exist
to facilitate testing and development to a single bootstrap script
instance which is shared.