Refactor MscmpSystAuthn.Password check_recent_pwd
Previously password validity checks required knowing the Access Account
to test a credential.  The reason for this was some validity checks
could depend on Owner Password Rules and getting those was done via the
Access Account association with Owners (or for determining the Access
Account was unowned).

However it can be desirable to test password against arbitrary rules,
especially the Global Password rule without necessarily knowing an
Access Account.  For example, Platform Bootstrapping wants to create the
first Owner and Access Account records; given this the creation process
either needs to be 2 step: create the Owner and Access Account and then
in the next step create and test the Access Account credential; or get
the Global Password Rules which are the only rules that can exist at
platform bootstrap time and just test the entered credential against the
default global rules.  The second option allows a single bootstrap step
for creating our records with credential and for reasonably testing the
credential against the Global Password Rules interactively at entry
time.  The changes here facilitate the second option.

The change here patches a function which would fail if there was no
provided Access Account; no Access Account means the question about if
it was using a recently used password is moot and so it should just
ignore the question.  Naturally there are contexts where we cannot just
assume that the Access Account doesn't matter; but whether or not to
care must be a concern of the caller anyway.