Add function for Generic Password Rules
There are structs describing Global Password Rules and Owner Password
Rules.  Insofar as actual rules definitions are concerned, these structs
are identical.  The function added here translates the specific kind
based Password Rule structs into a simple map which just keeps the
Password Rules values themselves.  This generic map of Password Values
can then be used to test passwords without caring whether or not the
Password Rule is the Owner Password Rules or the Global Password Rules.

Keep in mind that the concept of generic Password Rules is not new with
this commit.  The Effective Password Rules any one Access Account holder
will experience may be a combination of Global and Owner Password Rules
and that mixed expression was already present in the type
`Types.password_rules()`.  This function just makes it easier to get
a value of that type on demand.