Add no_bootstrap pipeline to MsappMcpWeb.Router
Previously, we put a routing pipeline to check if a call to the root of
the administrative endpoint should redirect to the bootstrap process if
the system state said we were bootstrapping.  This new check does the
reverse.  If someone tries to get to the bootstrapping form and we're
not bootstrapping, redirect to the root path of the admin endpoint for
login.