Add access_account_exists?/1 for MCP Bootstrap sup
MCP Platform bootstrapping shouldn't be necessary if Access Accounts
are already present; indeed, the whole point of the MCP bootstrapping is
to create an initial Access Account with which to perform further setup
and configuration.  This commit adds the API with which to test for the
presence of pre-existing Access Accounts.