Fix bad MscmpSystInstance.owner_exists?/1 returns
The return type was poorly conceived; the code was working as designed,
poorly.

Now actually match the conventions for a `?` ending function name.  The
only place where this now goes off the ranch is in sending an error
tuple in a spot where we don't otherwise send a success tuple.  This is
yet another unique way of returning in the code base.  Don't know if it
matters, certainly not the worst of sins; but the inconsistency doesn't
feel quite right either.