Move msform_data from MsappMcpWeb to MsappMcp
After a fair amount of thinking about where certain kinds of code should
live, I've landed on something that feels reasonable.

MsappMcp in a classic Phoenix umbrella application would be where the
business logic would be.  In our case, we don't need it for that since
our business logic should be in our components and subsystems when
aggregate actions need to be taken.

Rather than just saying that we won't use MsappMcp, I think there's a
narrow gap of functionality that isn't quite business logic, but also
isn't quite direct UI code like the components and templates are.  In
this space, for example the data definitions and processing of form
input.  We're still tightly coupled to UI flows and concerns, but we're
not talking logic used to directly generate HTML.  In the form
processing example, the form data could originate from a web page or a
public API call... we'd still want to validate it either way prior to
passing it to the actual business logic which will focus on data record
validation rather than UI input validation.

To this end move the MsformData module code from MsappMcpWeb to
MsappMcp.

Also, Rename MsformData to MsdataApi. This is to remove the notion that
these Embedded Schemas and validators are only for UI forms.  They are
for UI forms and External API data validation so call them more
generically.  In this sense the forms and the external API front end
have as goal completing validated API Data.