Add MssubMcp tenant bootstrapping
This general functionality is the start of MCP tenant bootstrapping.
The first specific functionality in this regard is actually
bootstrapping the platform; that looks a lot like tenant bootstrapping
except that we don't create an application instance at the end of the
process and administrative MCP permissions are assigned during the
platform bootstrapping process.

Note that the integration testing includes a permissions related API
test not captured earlier.