Update Msplatform bootstrap through record create
At this point we can get through the whole bootstrapping process and get
records created and platform_state updated to an active state.  The form
also redirects to a login form, though that form doesn't exist yet.

A fair number of changes in CoreComponents and in how certain tasks
happen.  In fact, we do add a Task.Supervisor since some longer running
processes need to run in their own process (e.g. disallowed password
loading) so that the live view socket doesn't get timed out and
restarted.

At this point, we're getting a sense of organization and what kinds of
things belong in MssubMcp and in MsappMcp (discussed in earlier commits)
as well as what components we'll need... but much is still in flux and
very much WIP in this commit.