Add Msplatform setting `platform_owner`
This change makes the assumption that anyone logging into the MCP for
administrative purposes is doing so using an Access Account owned by the
platform owner.  This means the Owner does not need to be specified at
login time and can be kept in a configuration point and applied without
demanding the user identify the owner in some way (directly, URL query
variables, etc).