Fix platform admin identity validation bug
When the system created the email identity for the Platform
Administrator, it was doing so as unvalidated, meaning that the admin
account couldn't be used to login and continue platform setup... such as
setting up an email server used to send out email identity validation
requests.

This fix causes the tenant setup process for application `:mcp` to
create the authenticator to use the `create_validated: true` option.

For now don't create regular tenants as validated as it might still be
appropriate to require validation for normal tenant setup.  We defer a
decision in that case for later.