Refactor MscmpSystAuthn verify_email_address
Prior to this commit.  verify_email_address/1 would return either the
submitted email address, if valid, or would raise an exception if the
email address was not considered valid.   In practice this doesn't work
so well.  A user may accidentally, or a malicious script may purposely,
submit an invalid email address and when encountered, we shouldn't have
to catch exceptions all the time.

So refactor MscmpSystAuthn.Impl.Identity.Email.verify_email_address/1 to
return a success/error tuple dependent on the validity of the email
address.

Also, if we're authenticating an Email/Password Authenticator, if the
starting email address is not valid, we should process the
authentication as being rejected from the get-go... we should also fake
the credential processing time to boot.