Complete MscmpSystSession
After some thinking and initial efforts to make this the direct
implementation of the Plug.Session.Store behaviour, it's clear that this
component is most appropriately not a strict implementation for web
sessions, but database backed sessions more generally.

The plan as of now is to included this component as a generic database
session store in MssubMcp and then to implement Plug.Session.Store at in
Msplatform/MsappMcp (or MsappMcpWeb) where we will already have Plug and
a clear need for a web front-end session store.

As of this commit, MscmpSystSession itself should be completed including
the remove of Plug & Cowboy as dependencies, the API, the documentation,
the unit and integration tests.  We should be ready to wrap this in
MssubMcp.