Refactor MscmpSystSession and dependents
After reading:

https://thepugautomatic.com/2020/05/persistent-session-data-via-a-database-in-phoenix-liveview/

it became clear that this was a superior approach to handling sessions
with LiveView, at least if you've already committed to storing session
data in the database anyway.  What I like about the approach is that it
decouples the HTTP request/response cycle from the actual mechanics of
the session which is a good match for something like LiveView.  This
means that the HTTP session data is reduced to those items required for
running the UI (like CSRF tokens) and a reference to the database record
which backs the session.

So the process now is that standard client/server session will managed
via the simple, out-of-box cookie session store.  On accessing the
login form we'll create a new "Session Name" and store that in the
session cookie... we'll get this on the login LiveViews mount.  If the
login is successful, we'll create the session record in the database
with the Authentication State (that the user was authenticated, the
instance to which they've been authenticated, etc.) and we'll use the
previously generated Session Name as the database record Session Name
(internal_name).  This should be OK so long as we trust that the out-of-
box cookie store does what it says on the tin and sign the cookie so
that any tampering isn't possible (otherwise, a user could pick an
arbitrary internal name rather than one we randomly generate).  All
checks for authentication or other session related data will then have
to get the session data from the database record for that session.  If
that session is deleted or expires, then the system will treat any
session related requests as though the session is not found and force
the start of a new session at login.

At the end of the day this change and the code changes enabled by this
seems much simpler than what we were going to try before.  My only
concern is the risk of having the session name transmitted to the client
and available in the cookie.  I don't think it's adding any additional
risk to being fully session based, but this new approach seems just to
easy and clean for something with security ramifications.