Rename MsdataApi to Msform
When the submodules of this parent module where expected to just be
Ecto embedded changesets, the MsdataApi name was reasonable.  But now
we're defining a fair bit more than than here and the "data" name
only describes some of the current functionality that may appear here.

Msform, thought of broadly enough to include any externally facing
interface including GUI forms and externally usable interfaces to our
data, encompasses more of what it is we're really defining here.