Add component MscmpSystForms
As we're getting more sophisticated with form building, time to get some
better defined functionality in place.  Ideally, we're be able to define
generic user interfaces here as well as being able to put together some
basic web components for the web user interface.  It will be interesting
to see if this is sufficiently resilient for actual use.

At this commit, we're at the scaffolded project phase with principally
only dependencies defined.