Add basic documentation site structure
"The Book" documentation site will be comprised of 4 relatively
independent sections: Theory, Planning, End User, Technical.   These
top level sections should reasonably partition information into audience
specific groups of topics.

This is far from complete, but it's a decent start on the idea of these
groupings and, on a much more practical note, the live reload function
of Hugo may need to have a git commit with time in order to function
correctly (currently my live reloading is not working).