Minor cleanup of McpBootstrap
This isn't the full clean-up needed (like adding docs and such) alluded
to in earlier commit messages, but some minor cleanup.

First, the Msforms structure which was going to split login into "Data",
"Definitions", and "Events" is basically still sound, but "Events" is
the wrong label for that file.  In this commit it's renamed to
"Actions".  "Events" are handled by the live controller, the live
controller will typically call a single dedicated "Action" function
which will then orchestrate the logic across business logic boundaries
as needed.

Second, remove a couple commented dead code leftovers and things like
that.

Tested that the form continues to work at least as well as it did (happy
path test).