Remove outdated comment from MscmpSystSession
The original comment was made after bringing in a lot of cruft to
support Plug and the idea was that an implementation closer to the UI
would implement the behaviour.

On further examination it doesn't seem necessary to wrap this module's
functionality at all.  I need the Plug.Session.COOKIE store which will
only hold the db's `session_name` value in a cookie.  So long as I can
get that across liveview's (and I should be able to without problem) the
database session can be completely managed with the MscmpSystSession API
directly.