Add password studies to MscmpSystAuthn
I wasn't happy with the "stregnth" note in the MscmpSystAuthn docs.  The
note itself was fine, but needed more to justify why I was making those
statements.

I include the statements which set the premise of my comments and also
include a newer study which provides a hybrid take that, to my reading,
is probably a better take on the question.  The only issue with
including the new study is that it recommend a strength evaluation at
entry time with suggestions and we don't implement those ideas here.

I have some hope that passkey like solutions may make this whole
password thing moot... we'll see how well those things stick.