Add MscmpSystSettings functions to MssubMcp API
As much as I hate it, we're still passing a lot of component APIs
through MssubMcp one-for-one; up to this point MscmpSystSettings was not
one of these.

This adds MscmpSystSettings API (not the whole API) exposure through
MssubMcp.  This is for consistency of usage and also because MssubMcp
does allow us to set the correct services for MCP access without having
to do so explicitly in all the highest level client code.