Add email password authentication
This is the first addition of Email/Password based authentication for
MCP.  This currently handles the following features:

  - Routing based on session authentication state
  - Login
  - Bad login attempt handling
  - User session retention of authentication status
  - Logout
  - Redirect to requested path on login

The functionality is complete enough to test logging in, logging out,
authentication related request routing, and sessions but should be
construed as complete.

Remaining features at this commit include:

  - Password recovery
  - Password validation
  - Password change/reset

And that list shouldn't be considered exhaustive.