Refactor msvalidated_button for focus issue
Previously, you could tab out of the last field of a form to an
msvalidated_button component.  The button component would take the
focus, but because the focus movement is faster than form validation,
the message button would take the focus before the button state had a
chance to update; this naturally disrupts the focus flow of the form
since when the message button gets hidden, the focused element becomes
essentially undefined.

Fix this by refactoring the msvalidated_button to use a single button
which changes its contents and attributes based on the current button
state rather than three buttons which are revealed or hidden based on
button state.  This way the button receiving focus is always the same
button regardless of state.