Refactor parameter type in MCP Msforms modules
With the AuthPasswordReset form, taking a look at how we're handling
types for use with typespecs led me to conclude we needed to have some
Msform related handling norms.

For Msform based forms, we want to define the type of the parameters
that can be passed into the Changeset generating functions and so we'll
have each Msform module define its own types which will usually just be
the `parameters` type for the module.  This will be in a `types.ex` file
kept along side the other Msform module specific files.

This commit refactors the existing Msform forms (McpBootstrap &
AuthEmailPassword) to use the new standard, removing any overlapping
vestigial code previously written for this purpose.

An additional task in this commit is also to bring McpBootstrap up to
standard by not using a Msform.McpBootstrap alias for internal
references and just accept that `Msform.` isn't enough to warrant the
alias... this will make McpBootstrap consistent will other modules now
avoiding that self-referencing alias.