Add reset_password_credential/2 to MscmpSystAuthn
The logic that could support password resets was build, but never
exposed.  This commit adds that exposure to the MscmpSystAuthn &
MssubMcp APIs.

This add includes documentation and tests where appropriate.   At this
commit all related tests in all components/subsystems are passing.  Also
there are no dialyzer or credo issues related to the changes here.