Refactor mssub_mcp DB migrations to mssub_mcp
Previously we had msplatform responsible for keeping the migrations for
subsystem mssub_mcp.  As I recall (I haven't looked at the commit notes
from days gone by), that we wanted to do this because things like the
MscmpSystInstance applications and applications context for mssub_bms
were loaded as part of the migrations.  mssub_mcp shouldn't have any
knowledge of mssub_bms, but the platform could rightfully have knowledge
of it.

In the end that rationale still didn't really make sense because unless
the platform specifically includes mssub_bms, it shouldn't be adding it
as an application.  In the end, mssub_bms will need to be responsible
for calling functionality to load its own application data if its part
of the build. (We delete the mssub_bms part of the migration outright
in this commit).

So with that out of the way, the clearly most sensible place for the
MCP migrations are in the mssub_mcp subsystem itself.  This became
evident and urgent as we moved the settings look-ups for the MCP
state and owner from msplatform, where we had to do extra context
setting in msplatform to be sure we were talking to the MCP services
when it could more easily and less repetitiously be added to mssub_mcp
API in a single place.  The settings were seed data in the platform and
adding the explicit setting lookup to mssub_mcp would have create a
back dependency of mssub_mcp on msplatform: which of course is not
acceptable.

So this commit effects the move of the migrations to mssub_mcp from
msplatform.  mssub_mcp testing, both unit and integrated, as well as
mssub_mcp focused credo and dialyzer runs all look good.  I have not
run these on msplatform as it's in pretty sad shape after the bootstrap,
login, and soon to be password reset work... so that will come later.