Relocate MCP settings lookup to mssub_mcp
Previously, using the session management functionality of mssub_mcp in
msplatform required a configuration setting
("platform_session_expiration") to be looked up in msplatform which also
meant that the PDict identified services had to also be updated at each
call.  This created a memory item each time we needed to work with the
session for this administrative chore.

Really the calls for the session management in mssub_mcp should already
have this expiration related formally handled.  First, it's handled in
a single set of functions that are colocated and only require that its
handled once... this way the many calls to mssub_mcp session management
functionality always also get and apply the setting without having to
keep remembering to do so.  Moreover, its just the correct level of
abstraction and the unnecessary repetition was a signal that the bad
code had erred in this way.

Move that functionality to mssub_mcp and make it flexible enough to
override the setting value if desired and update documentation to
reflect where the default value comes from.