Add MscmpSystAuthn struct Jason.Encoder derives
Downstream (in MsappMcp for now) there are attempts to put some structs
in the database session's data field and this can't happen without an
encoding to JSON.  Structs are not automatically able to be encoded by
the Jason library and so must be set up to be.

In truth, adding this encoding to the struct in the MscmpSystAuthn
component doesn't feel quite right.  There's nothing about
MscmpSystAuthn that needs this nor is it necessarily needed by any
consumer.  So doing the protocol implementation in MscmpSystAuthn seems
sort of like tightening the coupling to a consumer application which
should really do its own protocol implementation if its needed.

There are two reasons to do it at the struct itself:

  1 - It's easier.  Using the Elixir `@derive` functionality, we don't
      have to worry about the actual implementation details and we can
      get (for now) a good enough representation from the stock Jason
      library implementation.

  2 - If each application that needs to have MscmpSystAuthn data in JSON
      form is fully decoupled from all other applications that need it,
      we could end up with multiple attempts to creation the protocol
      implementation if these applications come together in the
      platform.  That wouldn't work.

Point 1 probably isn't enough to sway me to use MscmpSystAuthn for the
protocol implementation, but point 2 virtually requires it.

I could also create a separate component for the sole purpose of
implementing the protocols (whenever a library like MscmpSystAuthn is
found to be a dependency), I think it has all the downsides of just
doing it in MscmpSystAuthn itself (maybe implement the protocol for
objects not needed in a given profile) while having more code to do it
with.

So for now, implement the protocol directly in the schema modules like
is done here and just do it anytime a reason comes up to have it.  If it
continues to feel more and more like a crime against nature, address it
once it becomes unbearable.