Add MsappMcp.authenticate mode opts
Previously, MscmpMcp.authenticate/3 would authenticate a login style
form params and create a database session if the authentication was
successful.  This worked for user session authentication, but was
insufficient for all authentication use cases.

There are authentication use cases where the purpose of authentication
is not to validate the user's session, but to simply re-confirm that the
user is the one using an authentication session rather than some
opportunistic passer-by.  In this case, we want to go through all the
same authentication steps except creating a new database session.

This change adds an optional `mode` param (defaulting to `session`)
which causes (:session)/suppresses(:confirm) the creation of a database
session on successful authentication.  This means the main function is
now MsappMcp.authenticate/4.