Fix failing MscmpSystDb Decimal range doctest
The original version of the doctest relied on `inspect/2` outputting the
result of the example call in a certain format.  Somewhere along the way
the output changed and this test fails, not because the example produces
a wrong result, but because the method to test that result was janky
from the start.

The real answer here is that the Decimal library gives its results via a
Decimal struct.  So rather than test on the visual representation of
the struct returned as an answer... just use the actual struct we expect
the function to return as a result.  While the form of that stuct is a
touch less obvious than the original representation of that struct in an
iex context, it's nonetheless the correct thing to compare against.