Fix MscmpSystMcpPerms test failure; missing config
The issue here was that the dependencies ultimately led to
MscmpSystLimiter and Hammer.  Currently Hammer wants a config to be set
up in the application config (config.exs) and this hadn't been don in
this case.

Adding the Hammer config resolves the testing issue, but my sense is
that 1) we shouldn't use config.exs for library configuration; and 2) my
recollection is that it isn't necessary... but I don't want to tackle
that just right now.  So we'll go ahead and add the config to just get
through things, but memorialize this issue (here) for later.