Fix MscmpSystAuthn API function doc format issue
I had an in-sentence-list of two items which the ExDoc parse interpreted
the second item as a numbered list item because it was at the start of
the line with the correct indentation for that (from the markdown
perspective).

I actually think this could be a bug in ExDoc or the markdown processing
library.  The indentation which made this look like a valid list item
was not there because I was indenting from the left for this purpose but
because the heredoc string started an indent or two from the far left
column.  If the markdown processing assumed, as the rest of the heredoc
does, that left most column is considered the column directly over the
first of the heredoc closing `"""` characters, then it wouldn't see this
as a list item (at least I don't think it would).

Not a big deal.  The list may even be better than the original
formatting so rather than fight this battle... concede it.